# ENH: Decrease wall time of ``_get_PW_matrices`` for ``BottomUp`` and ``BottomUpSparse``