--- a/nbs/methods.ipynb
+++ b/nbs/methods.ipynb
@@ -326,10 +326,11 @@
     "\n",
     "    def _get_PW_matrices(self, S, idx_bottom):\n",
     "        n_hiers, n_bottom = S.shape\n",
-    "        P = sparse.lil_matrix(S.shape, dtype=np.float32)\n",
-    "        P[idx_bottom] = S[idx_bottom]\n",
-    "        P = sparse.csr_matrix(P.T)\n",
-    "        W = sparse.identity(n_hiers, dtype=np.float32)\n",
+    "        P = sparse.eye(n_bottom, n_hiers, n_hiers - n_bottom, np.float32, \"csr\")\n",
+    "        if getattr(self, \"intervals_method\", False) is None:\n",
+    "            W = None\n",
+    "        else:\n",
+    "            W = sparse.eye(n_hiers, dtype=np.float32, format=\"csr\")\n",
     "        return P, W"
    ]
   },