--- a/nbs/methods.ipynb
+++ b/nbs/methods.ipynb
@@ -887,7 +887,7 @@
     "        idxs_bu = np.hstack(idxs_bu)\n",
     "        #bottom up forecasts\n",
     "        bu = BottomUp().fit_predict(\n",
-    "            S=np.unique(S[idxs_bu], axis=1), \n",
+    "            S=np.fliplr(np.unique(S[idxs_bu], axis=1)), \n",
     "            y_hat=y_hat[idxs_bu], \n",
     "            idx_bottom=np.arange(len(idxs_bu))[-len(cut_nodes):]\n",
     "        )\n",